refactor: Moved to dto package
