--- src/main/java/com/wanted/preonboarding/recruit/RecruitDetailDto.java
@@ -1,5 +1,0 @@
-package com.wanted.preonboarding.recruit;
-
-//채용공고 상세 데이터 dto
-public class RecruitDetailDto {
-}
