--- src/main/java/com/wanted/preonboarding/recruit/RecruitResponseDto.java
@@ -1,5 +1,0 @@
-package com.wanted.preonboarding.recruit;
-
-//채용공고 조회 dto
-public class RecruitResponseDto {
-}
