+++ src/main/java/com/wanted/preonboarding/recruit/dto/RecruitDetailResponseDto.java
@@ -1,0 +1,5 @@
+package com.wanted.preonboarding.recruit.dto;
+
+//채용공고 상세 데이터 dto
+public class RecruitDetailResponseDto {
+}
