+++ src/main/java/com/wanted/preonboarding/recruit/dto/RecruitRegisterDto.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit;
+package com.wanted.preonboarding.recruit.dto;
 
 //채용공고 등록 데이터 dto
 public class RecruitRegisterDto {
