+++ src/main/java/com/wanted/preonboarding/recruit/dto/RecruitResponseDto.java
@@ -1,0 +1,6 @@
+package com.wanted.preonboarding.recruit.dto;
+
+//채용공고 조회 dto
+//채용공괴 조회 및 검색시 사용
+public class RecruitResponseDto {
+}
