+++ src/main/java/com/wanted/preonboarding/recruit/dto/RecruitUpdateDto.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit;
+package com.wanted.preonboarding.recruit.dto;
 
 //채용공고 수정 dto
 public class RecruitUpdateDto {
